CI: Move main setup code into common framework
