+++ test/common_framework.py
@@ -1,3 +1,4 @@
+import inspect
 import pexpect
 import string
 import random
@@ -14,7 +15,7 @@ from ptyprocess import PtyProcessError
 from shutil import copy, rmtree
 from subprocess import (Popen, call, check_call, check_output,
                         DEVNULL, STDOUT, TimeoutExpired, CalledProcessError)
-from sys import exit, stdout, stderr, version_info
+from sys import argv, exit, modules, stdout, stderr, version_info
 from tarfile import open as topen
 from time import sleep
 from unittest.util import strclass
@@ -669,3 +670,79 @@ def main(argv=None):
         exit("Python 3.2 or later is required.")
     failfast = not (environ.get("NO_FAILFAST"))
     unittest.main(testRunner=MyTestRunner, argv=argv, verbosity=2, failfast=failfast)
+
+
+def main_setup(testcase):
+
+    tests = [t[0] for t in
+            inspect.getmembers(testcase, predicate=inspect.isfunction)
+            if t[0].startswith("test")]
+
+    xtests = [t[0] for t in
+            inspect.getmembers(testcase, predicate=inspect.isfunction)
+            if t[0].startswith("xtest")]
+
+    cases = [c[0] for c in
+            inspect.getmembers(modules['__main__'], predicate=inspect.isclass)
+            if issubclass(c[1], testcase) and c[0] != testcase.__name__]
+
+    attrs = sorted(testcase.attrs)
+
+    def explode(n, attr=None):
+        if n in tests:
+            return [c + "." + n for c in cases]
+        if n in xtests:
+            return [c + "." + n for c in cases]
+        if n in cases:
+            if attr:
+                return [n + "." + t[0] for t in
+                    inspect.getmembers(testcase, predicate=inspect.isfunction)
+                    if hasattr(t[1], attr)]
+            else:
+                return [n,]
+        p = n.split('.')
+        if p and p[0] in cases and (p[1] in tests or p[1] in xtests):
+            return [n,]
+        return []
+
+    if len(argv) > 1:
+        if argv[1] == "--help":
+            print(("Usage: %s [--help | --get-test-binaries | " +
+                   "--list-attrs | --list-cases | --list-tests] | " +
+                   "[--require-attr=STRING TestCase ...] | " +
+                   "[TestCase[.testname] ...]") % argv[0])
+            exit(0)
+        elif argv[1] == "--get-test-binaries":
+            get_test_binaries()
+            exit(0)
+        elif argv[1] == "--list-attrs":
+            for a in attrs:
+                print(str(a))
+            exit(0)
+        elif argv[1] == "--list-cases":
+            for m in cases:
+                print(str(m))
+            exit(0)
+        elif argv[1] == "--list-tests":
+            for m in tests:
+                print(str(m))
+            exit(0)
+        else:
+            x = re.match(r"^--require-attr=(\S+).*$", argv[1])
+            if x:
+                attr = x.groups()[0]
+                del argv[1]
+            else:
+                attr = None
+
+            a = []
+            for b in [explode(x, attr=attr) for x in argv[1:]]:
+                a.extend(b)
+
+            if not len(a):
+                print("No tests found, was your testcase or testname incorrect? See --help")
+                exit(1)
+            return [argv[0],] + a
+
+    return None
+
